[backend] Change file for utils function
